tools/weather-tool/requirements.txt
# Weather Tool dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-consul==1.1.0
requests==2.31.0

